# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Tag": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Memory": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"summary": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"summary",
					"type",
					"description",
					"data",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Memory.Exclude_keyofMemory.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"name",
					"summary",
					"type",
					"data"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Memory.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_Memory.Exclude_keyofMemory.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"MemoryWithTags": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"summary": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"summary",
					"type",
					"description",
					"data",
					"createdAt",
					"updatedAt",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemoryFocusRule": {
				"properties": {
					"id": {
						"type": "string"
					},
					"assistantId": {
						"type": "string"
					},
					"maxResults": {
						"type": "number",
						"format": "double"
					},
					"relationshipTypes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"priorityTags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"assistantId",
					"maxResults",
					"relationshipTypes",
					"priorityTags",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Assistant": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"chat",
							"assistant"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"type",
					"model",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_Assistant_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"assistant",
							"chat"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "gpt-assistants-api",
		"version": "0.2.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/memory": {
			"get": {
				"operationId": "GetMemories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Memory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Memory"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMemory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Memory"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Memory.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			}
		},
		"/memory/{id}": {
			"get": {
				"operationId": "GetMemory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Memory"
								}
							}
						}
					}
				},
				"tags": [
					"Memory"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMemory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Memory"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Memory"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMemory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Memory"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-owned/assistant/{assistantId}": {
			"get": {
				"operationId": "GetMemoriesByAssistantId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOwnedMemories",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memoryIds": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"memoryIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/memory-owned/owned/{assistantId}": {
			"get": {
				"operationId": "GetOwnedMemories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-owned/assistant/{assistantId}/memory/{memoryId}": {
			"post": {
				"operationId": "AddOwnedMemory",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveOwnedMemory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-focused/assistant/{assistantId}": {
			"get": {
				"operationId": "GetFocusedMemoriesByAssistantId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-focused/focus/{memoryFocusId}": {
			"get": {
				"operationId": "GetFocusedMemories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryFocusId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateFocusedMemories",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryFocusId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memoryIds": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"memoryIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/memory-focused/focus/{memoryFocusId}/memory/{memoryId}": {
			"post": {
				"operationId": "AddFocusedMemory",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryFocusId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveFocusedMemory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryFocusId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-focus-rule": {
			"post": {
				"operationId": "CreateMemoryFocusRule",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MemoryFocusRule"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"priorityTags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"relationshipTypes": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"maxResults": {
										"type": "number",
										"format": "double"
									},
									"assistantId": {
										"type": "string"
									}
								},
								"required": [
									"priorityTags",
									"relationshipTypes",
									"maxResults",
									"assistantId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/memory-focus-rule/assistant/{assistantId}": {
			"get": {
				"operationId": "GetMemoryFocusRules",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MemoryFocusRule"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-focus-rule/{ruleId}": {
			"get": {
				"operationId": "GetMemoryFocusRuleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MemoryFocusRule"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ruleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memory-focus-rule/{id}": {
			"put": {
				"operationId": "UpdateMemoryFocusRule",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemoryFocusRule"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveMemoryFocusRule",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/assistant/{id}": {
			"get": {
				"operationId": "GetAssistantById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Assistant"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Assistant_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}